+++ nzbhydra/searchmodules/newznab.py
@@ -522,7 +522,7 @@ class NewzNab(SearchModule):
         entry = self.create_nzb_search_result()
         # These are the values that absolutely must be contained in the response
         entry.title = item.find("title").text
-        if entry.title and "nzbgeek" in self.settings.host:
+        if entry.title and "nzbgeek" in self.settings.host and config.settings.searching.removeObfuscated:
             entry.title = entry.title.replace("-Obfuscated", "")
         if config.settings.searching.removeLanguage:
             for word in [" English", " Korean", " Spanish", " French", " German", " Italian", " Danish", " Dutch", " Japanese", " Cantonese", " Mandarin", " Russian", " Polish", " Vietnamese", " Swedish", " Norwegian", " Finnish", " Turkish", " Portuguese", " Flemish", " Greek", " Hungarian"]: