+++ static/js/nzbhydra.js
@@ -5580,6 +5580,14 @@ function ConfigFields($injector) {
                             }
                         },
                         {
+                            key: 'removeObfuscated',
+                            type: 'horizontalSwitch',
+                            templateOptions: {
+                                type: 'switch',
+                                label: 'Remove "obfuscated" from nzbgeek titles'
+                            }
+                        },
+                        {
                             key: 'nzbAccessType',
                             type: 'horizontalSelect',
                             templateOptions: {